+++ src/config.rs
@@ -57,9 +57,9 @@ use crate::config_http::{
 pub use crate::config_load_balance::{
     LoadBalanceConfig, LoadBalanceHealthCheckConfig, LoadBalanceHealthCheckExpectedHeader,
     LoadBalanceHealthCheckExpectedStatusRange, LoadBalanceHealthCheckProtocol,
-    LoadBalancePassiveHealthConfig, LoadBalancePersistenceConfig, LoadBalancePersistenceMode,
-    LoadBalanceQueueConfig, LoadBalanceRetryConfig, LoadBalanceSelection,
-    LoadBalanceSlowStartConfig,
+    LoadBalanceManagedCookieSameSite, LoadBalancePassiveHealthConfig, LoadBalancePersistenceConfig,
+    LoadBalancePersistenceMode, LoadBalanceQueueConfig, LoadBalanceRetryConfig,
+    LoadBalanceSelection, LoadBalanceSlowStartConfig,
 };
 pub use crate::config_loader::ConfigLoadError;
 #[cfg(feature = "load-balancer")]